Admin page. Popular events

+++ src/main/java/servlet/DeleteEventServlet.java
@@ -1,6 +1,11 @@
 package servlet;
 
 import dto.EventDto;
+import dto.ProfileDto;
+import dto.UserDto;
+import entity.Profile;
+import entity.User;
+import service.AuthenticationService;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -16,10 +21,17 @@ public class DeleteEventServlet extends HttpServlet {
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
 
         EventDto eventDto = EventDto.getInstance(req.getSession());
+        UserDto userDto = UserDto.getInstance(req.getSession());
+
+        Long userId = (Long) req.getSession().getAttribute(AuthenticationService.USER_AUTHENTICATION_KEY);
 
         try {
 
             Long id = Long.parseLong(req.getParameter("id"));
+
+            if (!userDto.isAdmin(userId) && !eventDto.findById(id).getOwner().getUserId().equals(userId)) {
+                throw new Exception("Invalid Permission");
+            }
 
             eventDto.deleteById(id);
 
@@ -27,6 +39,6 @@ public class DeleteEventServlet extends HttpServlet {
 
         String url = req.getHeader("referer");
 
-        resp.sendRedirect(url != null? url : "/EventsHere/my");
+        resp.sendRedirect(url != null? url : req.getContextPath() + "/my");
     }
 }
